aemet: rename weather_update_coordinator
Use "coordinator" instead, like other integrations.
