+++ homeassistant/components/aemet/coordinator.py
