+++ homeassistant/components/aemet/entity.py
@@ -8,7 +8,7 @@ from aemet_opendata.helpers import dict_nested_value
 from homeassistant.components.weather import Forecast
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from .weather_update_coordinator import WeatherUpdateCoordinator
+from .coordinator import WeatherUpdateCoordinator
 
 
 class AemetEntity(CoordinatorEntity[WeatherUpdateCoordinator]):
